refactor: move out from vue file job offers info
