+++ pages/empleo/ofertas.ts
@@ -1,0 +1,40 @@
+const somos_nlp_data_scientist = {
+    ocupación: 'Data Scientist',
+    "palabras clave": 'traducción',
+    contrato: 'T. Completo',
+    entidad: 'Somos NLP',
+    "nº trabajadores/as": '0 - 9',
+    localización: 'Remoto 100%',
+    "fecha publicación": '2022/09/09',
+    más: 'https://somosnlp.org'
+}
+
+const somos_nlp_nlp_engineer = {
+    ocupación: 'NLP Engineer',
+    "palabras clave": 'traducción',
+    contrato: 'T. Parcial',
+    entidad: 'Somos NLP',
+    "nº trabajadores/as": '10 - 49',
+    localización: 'Presencial Madrid',
+    "fecha publicación": '2022/10/01',
+    más: 'https://somosnlp.org/blog'
+}
+
+const somos_nlp_volunteer = {
+    ocupación: 'Volunteer',
+    "palabras clave": 'traducción',
+    contrato: 'Prácticas',
+    entidad: 'Somos NLP',
+    "nº trabajadores/as": 'Más de 250',
+    localización: 'Híbrido Alicante',
+    "fecha publicación": '2022/07/01',
+    más: 'https://somosnlp.org/hackathon'
+}
+
+export const columnas = ['ocupación', 'contrato', 'entidad', 'nº trabajadores/as', 'localización', 'fecha publicación', 'más']
+
+export const ofertas = [
+    somos_nlp_data_scientist,
+    somos_nlp_nlp_engineer,
+    somos_nlp_volunteer,
+]
